@@ -1,15 +1,28 @@
 <?php
 
-// app/Http/Controllers/HomeController.php
-
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
 class HomeController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
+     * Show the application dashboard.
+     *
+     * @return \Illuminate\Contracts\Support\Renderable
+     */
     public function index()
     {
-        return view('admin.home'); // Sesuaikan dengan path view
+        return view('home');
     }
 }
