@@ -2,39 +2,47 @@
 
 namespace Database\Seeders;
 
+use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\DB;
+use App\Models\User;
 
 class UserSeeder extends Seeder
 {
+    /**
+     * Run the database seeds.
+     */
     public function run(): void
     {
-        DB::table('users')->insert([
-            [
-                'name' => 'Admin',
-                'email' => 'admin@example.com',
-                'password' => Hash::make('password'), // Pastikan mengganti password sesuai kebutuhan
-                'role' => 'admin',
-                'created_at' => now(),
-                'updated_at' => now(),
-            ],
-            [
-                'name' => 'Staff Keuangan',
-                'email' => 'staffkeu@example.com',
-                'password' => Hash::make('password'),
-                'role' => 'staffkeu',
-                'created_at' => now(),
-                'updated_at' => now(),
-            ],
-            [
-                'name' => 'Staff Umum',
-                'email' => 'staffumum@example.com',
-                'password' => Hash::make('password'),
-                'role' => 'staffumum',
-                'created_at' => now(),
-                'updated_at' => now(),
-            ],
+        $admin = User::create([
+            'name' => 'Admin',
+            'email' => 'admin@example.com',
+            'password' => bcrypt('password'),
         ]);
+
+        $admin->assignRole('admin');
+
+        $staffkeu = User::create([
+            'name' => 'Staff Keu',
+            'email' => 'staffkeu@example.com',
+            'password' => bcrypt('password'),
+        ]);
+
+        $staffkeu->assignRole('staffkeu');
+
+        $staffumum = User::create([
+            'name' => 'Staff Umum',
+            'email' => 'staffumum@example.com',
+            'password' => bcrypt('password'),
+        ]);
+
+        $staffumum->assignRole('staffumum');
+
+        $user = User::create([
+            'name' => 'User',
+            'email' => 'user@example.com',
+            'password' => bcrypt('password'),
+        ]);
+
+        $user->assignRole('user');
     }
 }
